requirements.txt
Django>=4.2
djangorestframework>=3.14
SQLAlchemy>=2.0
